requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests
pydantic[email]
python-multipart
openai
authlib
itsdangerous
httpx
